Divided middlewares in three files
